增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用

+++ routes/web.php
@@ -35,15 +35,15 @@ Route::middleware(['auth'])->group(function () {
     // 修改單一球員表單
     Route::get('players/{id}/edit', [PlayersController::class, 'edit'])->where('id', '[0-9]+')->name('players.edit');
     // 刪除單一球員資料
-    Route::delete('players/delete/{id}', [PlayersController::class, 'destroy'])->where('id', '[0-9]+')->name('players.destroy');
+    Route::delete('players/delete/{id}', [PlayersController::class, 'destroy'])->where('id', '[0-9]+')->name('players.destroy')->middleware('can:admin');
     // 新增球員表單
-    Route::get('players/create', [PlayersController::class, 'create'])->name('players.create');
+    Route::get('players/create', [PlayersController::class, 'create'])->name('players.create')->middleware('can:admin');
     // 修改球員表單
     Route::get('players/{id}/edit', [PlayersController::class, 'edit'])->where('id', '[0-9]+')->name('players.edit');
     // 修改球員資料
     Route::patch('players/update/{id}', [PlayersController::class, 'update'])->where('id', '[0-9]+')->name('players.update');
     // 儲存新球員資料
-    Route::post('players/store', [PlayersController::class, 'store'])->where('id', '[0-9]+')->name('players.store');
+    Route::post('players/store', [PlayersController::class, 'store'])->where('id', '[0-9]+')->name('players.store')->middleware('can:admin');
 
     // 顯示顯示所有球隊資料
     Route::get('teams', [TeamsController::class, 'index'])->name('teams.index');
@@ -56,15 +56,15 @@ Route::middleware(['auth'])->group(function () {
     // 修改單一球隊表單
     Route::get('teams/{id}/edit', [TeamsController::class, 'edit'])->where('id', '[0-9]+')->name('teams.edit');
     // 刪除單一球隊及旗下球員資料
-    Route::delete('teams/delete/{id}', [TeamsController::class, 'destroy'])->where('id', '[0-9]+')->name('teams.destroy');
+    Route::delete('teams/delete/{id}', [TeamsController::class, 'destroy'])->where('id', '[0-9]+')->name('teams.destroy')->middleware('can:admin');
     // 新增球隊表單
-    Route::get('teams/create', [TeamsController::class, 'create'])->name('teams.create');
+    Route::get('teams/create', [TeamsController::class, 'create'])->name('teams.create')->middleware('can:admin');
     // 修改球隊表單
     Route::get('teams/{id}/edit', [TeamsController::class, 'edit'])->where('id', '[0-9]+')->name('teams.edit');
     // 修改球隊資料
     Route::patch('teams/update/{id}', [TeamsController::class, 'update'])->where('id', '[0-9]+')->name('teams.update');
     // 儲存新球隊資料
-    Route::post('teams/store', [TeamsController::class, 'store'])->name('teams.store');
+    Route::post('teams/store', [TeamsController::class, 'store'])->name('teams.store')->middleware('can:admin');
 });
